Add TBD markers
We are currently not sure if it makes sens to still run the benchmarks,
since none of the proposed solutions actually conform to the customers
expectations. I want to make clear in this commit, which parts are still
missing, so we can see what's missing when we continue working on this.